separated path naming test
